ServiceOwner Contracts split into different files
