+++ src/WebApiClient/Digdir.Library.Dialogporten.WebApiClient.ServiceOwner/Features/V1/Contracts/Action/DialogApiAction.cs
@@ -1,0 +1,46 @@
+using System.Text.Json.Serialization;
+
+namespace Altinn.ApiClients.Dialogporten.ServiceOwner.Features.V1.Contracts.Action;
+
+public class DialogApiAction
+{
+    /// <summary>
+    /// The unique identifier for the action in UUIDv7 format.
+    /// </summary>
+    [JsonPropertyName("id")]
+    public Guid Id { get; set; }
+
+    /// <summary>
+    /// String identifier for the action, corresponding to the "action" attributeId used in the XACML service policy,
+    /// <br/>which by default is the policy belonging to the service referred to by "serviceResource" in the dialog.
+    /// </summary>
+    [JsonPropertyName("action")]
+    public string Action { get; set; } = null!;
+
+    /// <summary>
+    /// Contains an authorization resource attributeId, that can used in custom authorization rules in the XACML service
+    /// <br/>policy, which by default is the policy belonging to the service referred to by "serviceResource" in the dialog.
+    /// <br/>
+    /// <br/>Can also be used to refer to other service policies.
+    /// </summary>
+    [JsonPropertyName("authorizationAttribute")]
+    public string? AuthorizationAttribute { get; set; }
+
+    /// <summary>
+    /// True if the authenticated user (set in the query) is authorized for this action.
+    /// </summary>
+    [JsonPropertyName("isAuthorized")]
+    public bool? IsAuthorized { get; set; }
+
+    /// <summary>
+    /// The logical name of the operation the API action refers to.
+    /// </summary>
+    [JsonPropertyName("name")]
+    public string? Name { get; set; }
+
+    /// <summary>
+    /// The endpoints associated with the action.
+    /// </summary>
+    [JsonPropertyName("endpoints")]
+    public ICollection<DialogApiActionEndpoint>? Endpoints { get; set; }
+}
