+++ src/WebApiClient/Digdir.Library.Dialogporten.WebApiClient.ServiceOwner/Features/V1/Contracts/Common/Links.cs
@@ -1,0 +1,9 @@
+using System.Text.Json.Serialization;
+
+namespace Altinn.ApiClients.Dialogporten.ServiceOwner.Features.V1.Contracts.Common;
+
+public class Links
+{
+    [JsonPropertyName("metadata")]
+    public string Metadata { get; set; } = null!;
+}
